Rename notification to notifications for unify format
