+++ notifications/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class NotificationConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'notification'
+    name = 'notifications'
