+++ notifications/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from notification.models import Notification
+from notifications.models import Notification
 
 
 class NotificationSerializer(serializers.ModelSerializer):
