+++ playpal_backend/settings.py
@@ -39,7 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'events',
-    'notification',
+    'notifications',
     'rest_framework_simplejwt',
     'accounts',
     'corsheaders'
